requirements.txt
asgiref==3.7.2
black==23.9.1
click==8.1.7
Django==4.2.5
django-cors-headers==4.2.0
django-dotenv==1.4.2
django-filter==23.3
djangorestframework==3.14.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
pytz==2023.3.post1
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.8.0
